Separated responsibilities into their own classes
Added FileHandler class to handle storing and reading images as well as deserializing json files. The Imgur API json parsing responsibility should probably be moved somewhere else in the future. Also, in the future, FileHandler should handle pngs as well as jpgs, possibly by automatically converting pngs to jpgs?

Added ImgurApiClient class to handle the actual API calls to the Imgur API. PostToImgur() now does return the response string and does not process it itself. The structure of the class is changed such that an HttpClient is only created once in the constructor and can be reused. In the future, this function should work asynchronous.